+++ Package/inst/config/scenarios/volta_wet_network.R
@@ -12,7 +12,6 @@ VoltaWetNetwork <- function(data_root, output_root) {
     enable_canals = TRUE,
     canal_shp_path = bc$canal_shp_path,
     canal_discharge_table = bc$canal_discharge_table,
-    canal_q_anchor_table = bc$canal_q_anchor_table,
     connect_canals_to_rivers = bc$connect_canals_to_rivers,
     canal_junction_snap_tolerance_m = bc$canal_junction_snap_tolerance_m,
     pop_raster_path = bc$pop_raster_path,